Reimplement positioning engine.
The previous relative 'grid' system was hard to reason about and error prone when jumping forward and back.
The new system is more explicit and way less complicated in its technical implementation.
An explicit 'positions' object is defined with key/value pairs denoting id and absolute position (not relative) of the item. This also drops the needs for 'insert' endpoint.