Reimplement item connections implementation.
Adding connections on a per item basis within the 'add' and 'update' action methods was getting unwieldy considering when changing or removing connections you have to use 'update'. The new implementation is modeled after the 'positions' API where a cascading dictionary is specified on each step that defines all connections.